pyproject.toml
[build-system]
requires = ["flit_core >=3.9.0,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "zfs-feature-discovery"
authors = [
    {name = "Daniel Miranda", email = "me@danielkza.io"},
]
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings",
    "aiofiles",
    "aioitertools",
    "argdantic",
    "pyyaml"
]
requires-python = ">=3.12"
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ['version', 'description']

[project.optional-dependencies]
test = [
    "ruff",
	"pytest<8.0.0",
    "pytest-mock",
    "pytest-asyncio",
]

[project.urls]
Documentation = "https://github.com/danielkza/zfs-feature-discovery"
Source = "https://github.com/danielkza/zfs-feature-discovery"
Changelog = "https://github.com/danielkza/zfs-feature-discovery/blob/develop/CHANGELOG.md"

[project.scripts]
zfs-feature-discovery = "zfs_feature_discovery.cli:main"
